+ Edit folder name

+++ src/containers/Folders/Folders.js
@@ -43,7 +43,7 @@ export class Folders extends Component {
 
     addFolder = () => {
         if (this.state.newFolderName === "") {
-            this.setState({isEditing: false});
+            this.setState({ isEditing: false });
             return;
         }
         const foldersToBeUpdated = [...this.state.folders];
@@ -63,25 +63,34 @@ export class Folders extends Component {
     }
 
     changeBeingEdittedFolderId = (folderId) => {
-        this.setState({idOfFolderBeingEditted:folderId})
-
+        if (folderId === this.state.idOfFolderBeingEditted) {
+            this.setState({ idOfFolderBeingEditted: -1 })
+        } else {
+            this.setState({ idOfFolderBeingEditted: folderId })
+        }
     }
-    // editFolder = (id) => {
-    //     const foldersToBeUpdated = [...this.state.folders];
-    //     const index = foldersToBeUpdated.findIndex(folder => folder.id === id);
-    //     foldersToBeUpdated[index]
-    // }
 
     deleteFolder = (id) => {
         const foldersToBeUpdated = [...this.state.folders];
         const index = foldersToBeUpdated.findIndex(folder => folder.id === id);
-        foldersToBeUpdated.splice(index, 1);
-        this.tasksElement.current.deleteFolder(id);
-        this.setState({ folders: foldersToBeUpdated });
+        if (window.confirm("Do you want to delete the folder '" + foldersToBeUpdated[index].name + "'?")) {
+            foldersToBeUpdated.splice(index, 1);
+            this.tasksElement.current.deleteFolder(id);
+            this.setState({ folders: foldersToBeUpdated });
+        }
     }
 
     updateCurrentFolder = (id) => {
         this.setState({ currentFolderId: id });
+    }
+
+    editFolderName = (e, id) => {
+        const foldersToBeUpdated = [...this.state.folders];
+        const index = foldersToBeUpdated.findIndex(folder => folder.id === id);
+        var newFolderName = e.target.value;
+        foldersToBeUpdated[index].name = newFolderName;
+        this.setState({ folders: foldersToBeUpdated });
+        console.log(this.state.folders)
     }
 
 
@@ -97,7 +106,8 @@ export class Folders extends Component {
                     key={eachFolder.id}
                     folder={eachFolder}
                     editFolder={this.editFolder}
-                    deleteFolder={this.deleteFolder}></Folder>
+                    deleteFolder={this.deleteFolder}
+                    editFolderName={this.editFolderName}></Folder>
             )
         });
         const plusSign = <FontAwesomeIcon
